refactor: 종료미션 더미 페이지 제거 & term page
