+++ src/pages/mypage/terms/terms.constants.ts
